multiple

+++ src/main/java/com/nbr/trp/inventory/controller/InventoryController.java
@@ -4,9 +4,13 @@ import com.nbr.trp.common.entity.LCStation;
 import com.nbr.trp.inventory.entity.Inventory;
 import com.nbr.trp.inventory.service.InventoryService;
 import com.nbr.trp.user.response.MessageResponse;
+import com.nbr.trp.user.service.UserDetailsImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -45,8 +49,17 @@ public class InventoryController {
     @GetMapping("/remove/{id}")
     public ResponseEntity<?> listInventory(@PathVariable String id) {
         try{
-            Boolean in = inventoryService.deleteInventory(id);
-            return ResponseEntity.ok(in);
+            SecurityContext context = SecurityContextHolder.getContext();
+            Authentication authentication = context.getAuthentication();
+            UserDetailsImpl userDetails1 = (UserDetailsImpl) authentication.getPrincipal();
+            String des = userDetails1.getDesignation();
+            if(des.equals("Director General")||des.equals("Assistant Programmer")) {
+                Boolean in = inventoryService.deleteInventory(id);
+                return ResponseEntity.ok(in);
+            }else{
+                return ResponseEntity.status(403).body("error");
+            }
+
         }catch(Exception e){
             return ResponseEntity.badRequest().body(new MessageResponse(e.getMessage()));
         }
